remove explicit csv handler and migrate tests to approp test module
